Add dcrinstall-manifests fake dist
This distribution does not perform any builds or archives, but instead
only generates the dcrinstall "manifest of manifests" for dcrinstall
to determine the latest software to fetch and install.  It requires
all other distributions to have already ran as it depends on the
manifests in the dist directory.

This replaces the compile_manifest.sh script from the decred-release
repo.