Use constants for decred and dexc relvers
This should make it easier to correctly update the release builder
with newer versions, under most situations.

Care will need to be taken if a new dexc version is released without a
new release of the other Decred tooling, as this will require using a
different const to point to a separate Github release for the new dexc
build.