Enable PIE on several targets
PIE provides additional hardening by loading the program code at
random offsets each execution, rather than all code being mapped to a
static memory location.  This complicates attacks such as ROP which
require jumping to known gadgets in the program text.

Other targets also support PIE, but it does not appear to be possible
to link the executable in a cross compile situation without the
platform's native linker.  This is the case for linux/386, linux/arm,
and freebsd/amd64.

As a consequence of producing a PIE, the Linux executables will now
link to the dynamic link editor (e.g. /lib64/ld-linux-x86-64.so.2 for
amd64), and will no longer run on distributions which do not provide
the link editor the expected path.  This is not expected to be a
common issue for most distributions, and for those that are affected
because of their reproducibility guarantees (e.g. NixOS), tooling is
available to patch the path to the dynamic link editor in the ELF
file.