Bump modules for v2.0.0 tag
This may not be strictly necessary since we did not design the builder to
contain any packages that other projects can import, but it seems like it
would be good form that we use a v2 module suffix for the v2.0.0 release tag.

Builds still match bit for bit after this change.