Prepare v2.1.3
Updates dcrd to release-v2.1.3 tag.

Updates dcrwallet to release-v2.1.3 tag.

Updates dcrctl to release-v2.1.3 tag.

Updates dcrdex mixing and wallet dependencies to v0.6.1 and v5.0.3
respectively, and modifies the version to v1.0.6 using the ldflags.  This will
closely match the build of the eventual v1.0.6 dcrdex tag, which won't be
available for some time.